-added memory allocator for buffer1D.h -added CPU capability checking (SSE, SSE2, MMX, 3DNOW) -added traits If<A, B, COND>

+++ devl/mvv/resource.h
@@ -8,7 +8,7 @@
 
 namespace mvv
 {
-   typedef nll::imaging::VolumeSpatial<double> MedicalVolume;
+   typedef nll::imaging::VolumeSpatial<float> MedicalVolume;
 
    class ResourceOrderList : public DynamicResource
    {
